cleanup: move adobe code to xds_adobe.go
